fixed workflow model
